sample settings component
